+++ tests/SortedSets/zPopMaxTest.php
@@ -3,9 +3,10 @@
 namespace Webdcg\Redis\Tests;
 
 use PHPUnit\Framework\TestCase;
+use Webdcg\Redis\Exceptions\SetOperationException;
 use Webdcg\Redis\Redis;
 
-class zRangeTest extends TestCase
+class zPopMaxTest extends TestCase
 {
     protected $redis;
     protected $key;
@@ -17,60 +18,18 @@ class zRangeTest extends TestCase
         $this->redis = new Redis();
         $this->redis->connect();
         $this->redis->setOption(\Redis::OPT_SERIALIZER, \Redis::SERIALIZER_NONE);
-        $this->key = 'SortedSets:zRange';
-        $this->keyOptional = 'SortedSets:zRange:Optional';
+        $this->key = 'SortedSets:zPopMax';
+        $this->keyOptional = 'SortedSets:zPopMax:Optional';
     }
 
     /*
      * ========================================================================
-     * zRange
+     * zPopMax
      * ========================================================================
      */
 
     /** @test */
-    public function redis_sorted_sets_zrange_with_scores()
-    {
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.1, 'A'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 2.2, 'B'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 3.3, 'C'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 3.3, 'D'));
-        $this->assertEquals(4, $this->redis->zCard($this->key));
-        // T E S T  -----------------------------------------------------------
-        $range = $this->redis->zRange($this->key, 1, 2, true);
-        $this->assertIsArray($range);
-        $this->assertEquals(2, count($range));
-        $this->assertArraySubset(['B' => 2.2], $range);
-        $this->assertArraySubset(['C' => 3.3], $range);
-        $this->assertArrayNotHasKey('A', $range);
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zrange_start_end()
-    {
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.1, 'A'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 2.2, 'B'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 3.3, 'C'));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 3.3, 'D'));
-        $this->assertEquals(4, $this->redis->zCard($this->key));
-        // T E S T  -----------------------------------------------------------
-        $range = $this->redis->zRange($this->key, 1, 2);
-        $this->assertIsArray($range);
-        $this->assertEquals(2, count($range));
-        $this->assertContains('B', $range);
-        $this->assertContains('C', $range);
-        $this->assertNotContains('A', $range);
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zrange_start()
+    public function redis_sorted_sets_zpop_max()
     {
         // Start from scratch
         $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
@@ -79,18 +38,17 @@ class zRangeTest extends TestCase
         $this->assertEquals(1, $this->redis->zAdd($this->key, 3.3, 'C'));
         $this->assertEquals(3, $this->redis->zCard($this->key));
         // T E S T  -----------------------------------------------------------
-        $range = $this->redis->zRange($this->key, 1);
-        $this->assertIsArray($range);
-        $this->assertEquals(2, count($range));
-        $this->assertContains('B', $range);
-        $this->assertContains('C', $range);
-        $this->assertNotContains('A', $range);
+        $max = $this->redis->zPop($this->key, 2, true);
+        $this->assertIsArray($max);
+        $this->assertArraySubset(['B' => 2.2], $max);
+        $this->assertArraySubset(['C' => 3.3], $max);
+        $this->assertEquals(1, $this->redis->zCard($this->key));
         // Remove all the keys used
         $this->assertEquals(1, $this->redis->delete($this->key));
     }
 
     /** @test */
-    public function redis_sorted_sets_zrange_defaults()
+    public function redis_sorted_sets_zpopmax_outbounds()
     {
         // Start from scratch
         $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
@@ -99,13 +57,49 @@ class zRangeTest extends TestCase
         $this->assertEquals(1, $this->redis->zAdd($this->key, 3.3, 'C'));
         $this->assertEquals(3, $this->redis->zCard($this->key));
         // T E S T  -----------------------------------------------------------
-        $range = $this->redis->zRange($this->key);
-        $this->assertIsArray($range);
-        $this->assertEquals(3, count($range));
-        $this->assertContains('A', $range);
-        $this->assertContains('B', $range);
-        $this->assertContains('C', $range);
-        $this->assertNotContains('D', $range);
+        $max = $this->redis->zPopMax($this->key, 5);
+        $this->assertIsArray($max);
+        $this->assertArraySubset(['A' => 1.1], $max);
+        $this->assertArraySubset(['B' => 2.2], $max);
+        $this->assertArraySubset(['C' => 3.3], $max);
+        $this->assertEquals(0, $this->redis->zCard($this->key));
+        // Remove all the keys used
+        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
+    }
+
+    /** @test */
+    public function redis_sorted_sets_zpopmax_multiple()
+    {
+        // Start from scratch
+        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
+        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.1, 'A'));
+        $this->assertEquals(1, $this->redis->zAdd($this->key, 2.2, 'B'));
+        $this->assertEquals(1, $this->redis->zAdd($this->key, 3.3, 'C'));
+        $this->assertEquals(3, $this->redis->zCard($this->key));
+        // T E S T  -----------------------------------------------------------
+        $max = $this->redis->zPopMax($this->key, 2);
+        $this->assertIsArray($max);
+        $this->assertArraySubset(['B' => 2.2], $max);
+        $this->assertArraySubset(['C' => 3.3], $max);
+        $this->assertEquals(1, $this->redis->zCard($this->key));
+        // Remove all the keys used
+        $this->assertEquals(1, $this->redis->delete($this->key));
+    }
+
+    /** @test */
+    public function redis_sorted_sets_zpopmax_default()
+    {
+        // Start from scratch
+        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
+        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.1, 'A'));
+        $this->assertEquals(1, $this->redis->zAdd($this->key, 2.2, 'B'));
+        $this->assertEquals(1, $this->redis->zAdd($this->key, 3.3, 'C'));
+        $this->assertEquals(3, $this->redis->zCard($this->key));
+        // T E S T  -----------------------------------------------------------
+        $max = $this->redis->zPopMax($this->key);
+        $this->assertIsArray($max);
+        $this->assertEquals(['C' => 3.3], $max);
+        $this->assertEquals(2, $this->redis->zCard($this->key));
         // Remove all the keys used
         $this->assertEquals(1, $this->redis->delete($this->key));
     }
